[sc-148974] cleanup
