+++ cmd/dedicated_inference/deployment.go
@@ -10,6 +10,7 @@ import (
 	exocmd "github.com/exoscale/cli/cmd"
 	"github.com/exoscale/cli/pkg/globalstate"
 	"github.com/exoscale/cli/pkg/output"
+	"github.com/exoscale/cli/utils"
 	v3 "github.com/exoscale/egoscale/v3"
 	"github.com/spf13/cobra"
 )
@@ -93,6 +94,25 @@ func (c *deploymentListCmd) CmdRun(_ *cobra.Command, _ []string) error {
 	return c.OutputFunc(&out, nil)
 }
 
+// resolveDeploymentID resolves a deployment UUID from an ID or a name.
+func resolveDeploymentID(ctx context.Context, client *v3.Client, nameOrID string) (v3.UUID, error) {
+	if id, err := v3.ParseUUID(nameOrID); err == nil {
+		return id, nil
+	}
+	resp, err := client.ListDeployments(ctx)
+	if err != nil {
+		var zero v3.UUID
+		return zero, err
+	}
+	for _, d := range resp.Deployments {
+		if d.Name == nameOrID {
+			return d.ID, nil
+		}
+	}
+	var zero v3.UUID
+	return zero, fmt.Errorf("deployment %q not found", nameOrID)
+}
+
 // create
 
 type deploymentCreateCmd struct {
@@ -153,7 +173,7 @@ func (c *deploymentCreateCmd) CmdRun(_ *cobra.Command, _ []string) error {
 		}
 	}
 
-	if err := runAsync(ctx, client, fmt.Sprintf("Creating deployment %q...", c.Name), func(ctx context.Context, c *v3.Client) (*v3.Operation, error) {
+	if err := utils.RunAsync(ctx, client, fmt.Sprintf("Creating deployment %q...", c.Name), func(ctx context.Context, c *v3.Client) (*v3.Operation, error) {
 		return c.CreateDeployment(ctx, req)
 	}); err != nil {
 		return err
@@ -196,7 +216,7 @@ func (c *deploymentDeleteCmd) CmdRun(_ *cobra.Command, _ []string) error {
 		return err
 	}
 
-	if err := runAsync(ctx, client, fmt.Sprintf("Deleting deployment %s...", c.Deployment), func(ctx context.Context, c *v3.Client) (*v3.Operation, error) {
+	if err := utils.RunAsync(ctx, client, fmt.Sprintf("Deleting deployment %s...", c.Deployment), func(ctx context.Context, c *v3.Client) (*v3.Operation, error) {
 		return c.DeleteDeployment(ctx, id)
 	}); err != nil {
 		return err
@@ -241,7 +261,7 @@ func (c *deploymentScaleCmd) CmdRun(_ *cobra.Command, _ []string) error {
 	}
 
 	req := v3.ScaleDeploymentRequest{Replicas: c.Size}
-	if err := runAsync(ctx, client, fmt.Sprintf("Scaling deployment %s to %d...", c.Deployment, c.Size), func(ctx context.Context, c *v3.Client) (*v3.Operation, error) {
+	if err := utils.RunAsync(ctx, client, fmt.Sprintf("Scaling deployment %s to %d...", c.Deployment, c.Size), func(ctx context.Context, c *v3.Client) (*v3.Operation, error) {
 		return c.ScaleDeployment(ctx, id, req)
 	}); err != nil {
 		return err
